refactor(auth): move logic handle from controller to service
