+++ be/shopveeAPI/Services/Auth/IAuthService.cs
@@ -1,4 +1,5 @@
 using Microsoft.AspNetCore.Mvc;
+using shopveeAPI.Services.Auth.Dto.Response;
 using shopveeAPI.Services.User.Dto.Request;
 
 namespace shopveeAPI.Services.Auth;
@@ -6,6 +7,5 @@ using Models;
 
 public interface IAuthService
 {
-    Task<User?> Login(UserRequest userRequest);
-    Task<IActionResult> UpdateRefreshToken(UserUpdateRefreshTokenRequest userUpdateRefreshTokenRequest);
+    Task<AuthResponse> Login(UserRequest userRequest);
 }
